feat: split TopologyMetrics into TopologyMetrics & TopologyManagerMetrics
