+++ zeebe/dynamic-config/src/main/java/io/camunda/zeebe/dynamic/config/ClusterConfigurationManagerService.java
@@ -27,6 +27,7 @@ import io.camunda.zeebe.dynamic.config.changes.PartitionChangeExecutor;
 import io.camunda.zeebe.dynamic.config.changes.PartitionScalingChangeExecutor;
 import io.camunda.zeebe.dynamic.config.gossip.ClusterConfigurationGossiper;
 import io.camunda.zeebe.dynamic.config.gossip.ClusterConfigurationGossiperConfig;
+import io.camunda.zeebe.dynamic.config.metrics.TopologyManagerMetrics;
 import io.camunda.zeebe.dynamic.config.metrics.TopologyMetrics;
 import io.camunda.zeebe.dynamic.config.serializer.ProtoBufSerializer;
 import io.camunda.zeebe.dynamic.config.state.ClusterConfiguration;
@@ -60,6 +61,7 @@ public final class ClusterConfigurationManagerService
   private final Actor managerActor;
   private final ClusterChangeExecutor clusterChangeExecutor;
   private final TopologyMetrics topologyMetrics;
+  private final TopologyManagerMetrics topologyManagerMetrics;
 
   public ClusterConfigurationManagerService(
       final Path dataRootDirectory,
@@ -72,6 +74,7 @@ public final class ClusterConfigurationManagerService
     this.clusterChangeExecutor = clusterChangeExecutor;
     this.memberShipService = memberShipService;
     topologyMetrics = new TopologyMetrics(meterRegistry);
+    topologyManagerMetrics = new TopologyManagerMetrics(meterRegistry);
     try {
       FileUtil.ensureDirectoryExists(dataRootDirectory);
     } catch (final IOException e) {
@@ -86,7 +89,7 @@ public final class ClusterConfigurationManagerService
     managerActor = new Actor() {};
     clusterConfigurationManager =
         new ClusterConfigurationManagerImpl(
-            managerActor, localMemberId, persistedClusterConfiguration, topologyMetrics);
+            managerActor, localMemberId, persistedClusterConfiguration, topologyManagerMetrics);
     clusterConfigurationGossiper =
         new ClusterConfigurationGossiper(
             gossipActor,
