+++ zeebe/dynamic-config/src/main/java/io/camunda/zeebe/dynamic/config/metrics/TopologyMetrics.java
@@ -12,19 +12,11 @@ import static io.camunda.zeebe.dynamic.config.metrics.TopologyMetricsDoc.*;
 import io.camunda.zeebe.dynamic.config.state.ClusterChangePlan;
 import io.camunda.zeebe.dynamic.config.state.ClusterChangePlan.Status;
 import io.camunda.zeebe.dynamic.config.state.ClusterConfiguration;
-import io.camunda.zeebe.dynamic.config.state.ClusterConfigurationChangeOperation;
 import io.camunda.zeebe.dynamic.config.state.CompletedChange;
-import io.camunda.zeebe.util.CloseableSilently;
-import io.camunda.zeebe.util.collection.Table;
 import io.camunda.zeebe.util.micrometer.EnumMeter;
-import io.camunda.zeebe.util.micrometer.MicrometerUtil;
-import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.MeterRegistry;
-import io.micrometer.core.instrument.Timer;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 
 public final class TopologyMetrics {
@@ -35,8 +27,6 @@ public final class TopologyMetrics {
   private final AtomicLong changeVersion;
   private final AtomicLong pendingOperations;
   private final AtomicLong completedOperations;
-  private final Map<String, Timer> operationDuration;
-  private final Table<String, String, Counter> operationAttempts;
   private final EnumMeter<Status> changeStatus;
 
   public TopologyMetrics(final MeterRegistry registry) {
@@ -52,27 +42,6 @@ public final class TopologyMetrics {
     changeVersion = makeGauge(CHANGE_VERSION);
     pendingOperations = makeGauge(PENDING_OPERATIONS);
     completedOperations = makeGauge(COMPLETED_OPERATIONS);
-    operationDuration = new HashMap<>();
-    operationAttempts = Table.simple();
-  }
-
-  private Timer registerOperation(final String operation) {
-    return Timer.builder(OPERATION_DURATION.getName())
-        .description(OPERATION_DURATION.getDescription())
-        .serviceLevelObjectives(OPERATION_DURATION.getTimerSLOs())
-        .tags(TopologyMetricsKeyName.OPERATION.asString(), operation)
-        .register(registry);
-  }
-
-  private Counter registerAttempt(final String operation, final String outcome) {
-    return Counter.builder(OPERATION_ATTEMPTS.getName())
-        .description(OPERATION_DURATION.getDescription())
-        .tags(
-            TopologyMetricsKeyName.OPERATION.asString(),
-            operation,
-            TopologyMetricsKeyName.OUTCOME.asString(),
-            outcome)
-        .register(registry);
   }
 
   private AtomicLong makeGauge(final TopologyMetricsDoc meter) {
@@ -105,47 +74,5 @@ public final class TopologyMetrics {
             .map(ClusterChangePlan::completedOperations)
             .map(List::size)
             .orElse(0));
-  }
-
-  public OperationObserver observeOperation(final ClusterConfigurationChangeOperation operation) {
-    return startOperation(operation);
-  }
-
-  OperationObserver startOperation(final ClusterConfigurationChangeOperation operation) {
-    final var timer =
-        operationDuration.computeIfAbsent(
-            operation.getClass().getSimpleName(), this::registerOperation);
-    return new OperationObserver(operation, MicrometerUtil.timer(timer, Timer.start(registry)));
-  }
-
-  public final class OperationObserver {
-    private final ClusterConfigurationChangeOperation operation;
-    private final CloseableSilently timer;
-
-    private OperationObserver(
-        final ClusterConfigurationChangeOperation operation, final CloseableSilently timer) {
-      this.operation = operation;
-      this.timer = timer;
-    }
-
-    public void failed() {
-      timer.close();
-      operationAttempts
-          .computeIfAbsent(
-              operation.getClass().getSimpleName(),
-              Outcome.FAILED.getName(),
-              TopologyMetrics.this::registerAttempt)
-          .increment();
-    }
-
-    public void applied() {
-      timer.close();
-      operationAttempts
-          .computeIfAbsent(
-              operation.getClass().getSimpleName(),
-              Outcome.APPLIED.getName(),
-              TopologyMetrics.this::registerAttempt)
-          .increment();
-    }
   }
 }
