+++ zeebe/dynamic-config/src/main/java/io/camunda/zeebe/dynamic/config/metrics/TopologyMetricsDoc.java
@@ -10,8 +10,6 @@ package io.camunda.zeebe.dynamic.config.metrics;
 import io.camunda.zeebe.util.micrometer.ExtendedMeterDocumentation;
 import io.micrometer.common.docs.KeyName;
 import io.micrometer.core.instrument.Meter.Type;
-import java.time.Duration;
-import java.util.stream.Stream;
 
 @SuppressWarnings("NullableProblems")
 public enum TopologyMetricsDoc implements ExtendedMeterDocumentation {
@@ -68,7 +66,7 @@ public enum TopologyMetricsDoc implements ExtendedMeterDocumentation {
 
     @Override
     public KeyName[] getKeyNames() {
-      return new KeyName[] {TopologyMetricsKeyName.CLUSTER_CHANGE_STATUS};
+      return TopologyMetricsKeyName.values();
     }
   },
   /** The version of the cluster topology change plan */
@@ -121,98 +119,15 @@ public enum TopologyMetricsDoc implements ExtendedMeterDocumentation {
     public String getDescription() {
       return "Number of completed changes in the current change plan";
     }
-  },
-  /** Duration it takes to apply an operation */
-  OPERATION_DURATION {
-    private static final Duration[] TIMER_SLOS =
-        Stream.of(
-                100, 1_000, 2_000, 5_000, 10_000, 30_000, 60_000, 120_000, 180_000, 300_000,
-                600_000)
-            .map(Duration::ofMillis)
-            .toArray(Duration[]::new);
-
-    @Override
-    public String getName() {
-      return "zeebe.cluster.changes.operation.duration";
-    }
-
-    @Override
-    public Type getType() {
-      return Type.TIMER;
-    }
-
-    @Override
-    public String getDescription() {
-      return "Duration it takes to apply an operation";
-    }
-
-    @Override
-    public Duration[] getTimerSLOs() {
-      return TIMER_SLOS;
-    }
-
-    @Override
-    public KeyName[] getKeyNames() {
-      return new KeyName[] {TopologyMetricsKeyName.OPERATION};
-    }
-  },
-  /** Number of attempts per operation type */
-  OPERATION_ATTEMPTS {
-    @Override
-    public String getName() {
-      return "zeebe.cluster.changes.operation.attempts";
-    }
-
-    @Override
-    public Type getType() {
-      return Type.COUNTER;
-    }
-
-    @Override
-    public String getDescription() {
-      return "Number of attempts per operation type";
-    }
-
-    @Override
-    public KeyName[] getKeyNames() {
-      return TopologyMetricsKeyName.values();
-    }
   };
 
   @SuppressWarnings("NullableProblems")
   public enum TopologyMetricsKeyName implements KeyName {
-    OPERATION {
-      @Override
-      public String asString() {
-        return "operation";
-      }
-    },
-    OUTCOME {
-      @Override
-      public String asString() {
-        return "outcome";
-      }
-    },
     CLUSTER_CHANGE_STATUS {
       @Override
       public String asString() {
         return "zeebe_cluster_changes_status";
       }
-    }
-  }
-
-  public enum Outcome {
-    FAILED("failed"),
-    APPLIED("applied");
-
-    private final String name;
-
-    Outcome(final String name) {
-      this.name = name;
-    }
-
-    public String getName() {
-      return name;
     }
   }
 }
